+++ baseline2/medal_contender/model.py
@@ -1,113 +1,140 @@
 import torch
-from transformers.modeling_outputs import TokenClassifierOutput
+import torch.nn as nn
+from torch.optim import lr_scheduler
+from transformers import AutoModel, AutoConfig
+from transformers import get_linear_schedule_with_warmup, get_cosine_schedule_with_warmup
+from medal_contender.configs import SCHEDULER_LIST, BERT_MODEL_LIST
 
+def init_params(module_lst):
+    for module in module_lst:
+        for param in module.parameters():
+            if param.dim() > 1:
+                torch.nn.init.xavier_uniform_(param)
+    return
 
-def __init__(self, config):
-    super(self.PreTrainedModel, self).__init__(config)
+class NBMEModel(torch.nn.Module):
+    def __init__(self, cfg, config_path=None, pretrained=False, model_type = 'base'):
+        super().__init__()
+        self.cfg = cfg
+        model_name = f"../models/{BERT_MODEL_LIST[cfg.model_param.model_name]}"
+        model_type = 'base' if 'base' in BERT_MODEL_LIST[cfg.model_param.model_name] else 'large'
+        if config_path is None:
+            self.config = AutoConfig.from_pretrained(
+                model_name)
+            self.config.update({"output_hidden_states": True,
+                       "hidden_dropout_prob": 0.2,
+                       "layer_norm_eps": 1e-7})
+        else:
+            self.config = torch.load(config_path)
 
-    kwargs = {
-        "add_pooling_layer": False
-    }
-    if config.model_type not in {"bert", "roberta"}:
-        kwargs = {}
-    setattr(self, self.backbone_name, self.ModelClass(config, **kwargs))
+        if pretrained:
+            self.backbone = AutoModel.from_pretrained(
+                model_name, config=self.config)
+        else:
+            self.model = AutoModel(self.config)
+        self.dropout = torch.nn.Dropout(p=0.2)
 
-    classifier_dropout_name = None
-    for key in dir(config):
-        if ("classifier" in key or "hidden" in key) and "dropout" in key:
-            if getattr(config, key) is not None:
-                classifier_dropout_name = key
-                break
+        if model_type == 'large':
+            self.classifier = torch.nn.Sequential(
+                    nn.Linear(self.config.hidden_size, 512),
+                    nn.LayerNorm(512),
+                    nn.Dropout(0.2),
+                    nn.ReLU(),
+                    nn.Linear(512, 1),
+                )
+        else:
+            self.classifier = torch.nn.Sequential(
+                    nn.Linear(self.config.hidden_size, 256),
+                    nn.LayerNorm(256),
+                    nn.Dropout(0.2),
+                    nn.ReLU(),
+                    nn.Linear(256, 1),
+                )
+        init_params([self.classifier])
+        
+    def forward(self, input_ids, attention_mask):
+        pooler_outputs = self.backbone(input_ids=input_ids, 
+                                       attention_mask=attention_mask)[0]
+        logits = self.classifier(self.dropout(pooler_outputs)).squeeze(-1)
+        return logits
 
-    if classifier_dropout_name is None:
-        raise ValueError("Cannot infer dropout name in config")
-    classifier_dropout = getattr(config, classifier_dropout_name)
-    self.dropout = torch.nn.Dropout(classifier_dropout)
-    self.classifier = torch.nn.Linear(config.hidden_size, 1)
+def fetch_scheduler(optimizer, cfg, num_train_steps=None):
+    '''
+        Config에 맞는 Solver Scheduler를 반환합니다.
+    '''
+    if cfg.model_param.scheduler=='linear':
+        scheduler = get_linear_schedule_with_warmup(
+            optimizer, num_warmup_steps=int(cfg.model_param.num_warmup_steps),
+            num_training_steps=num_train_steps
+        )
+    elif cfg.model_param.scheduler=='cosine':
+        scheduler = get_cosine_schedule_with_warmup(
+            optimizer, num_warmup_steps=int(cfg.model_param.num_warmup_steps),
+            num_training_steps=num_train_steps,
+            num_cycles=float(cfg.model_param.num_cycles)
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'CosineAnnealingLR':
+        scheduler = lr_scheduler.CosineAnnealingLR(
+            optimizer,
+            T_max=cfg.train_param.T_max,
+            eta_min=float(cfg.train_param.min_lr)
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'CosineAnnealingWarmRestarts':
+        scheduler = lr_scheduler.CosineAnnealingWarmRestarts(
+            optimizer,
+            T_0=cfg.train_param.T_0,
+            eta_min=float(cfg.train_param.min_lr)
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'LambdaLR':
+        scheduler = lr_scheduler.LambdaLR(
+            optimizer,
+            lr_lambda=lambda epoch: cfg.train_param.reduce_ratio ** epoch
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'MultiplicativeLR':
+        scheduler = lr_scheduler.MultiplicativeLR(
+            optimizer,
+            lr_lambda=lambda epoch: cfg.train_param.reduce_ratio ** epoch
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'StepLR':
+        scheduler = lr_scheduler.StepLR(
+            optimizer,
+            step_size=cfg.train_param.step_size, gamma=cfg.train_param.gamma
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'MultiStepLR':
+        scheduler = lr_scheduler.MultiStepLR(
+            optimizer,
+            milestones=cfg.train_param.milestones, gamma=cfg.train_param.gamma
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'ExponentialLR':
+        scheduler = lr_scheduler.ExponentialLR(
+            optimizer,
+            gamma=cfg.train_param.gamma
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'ReduceLROnPlateau':
+        scheduler = lr_scheduler.ReduceLROnPlateau(
+            optimizer,
+            mode='min',
+            patience=2,
+            min_lr=float(cfg.train_param.min_lr)
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'CyclicLR':
+        scheduler = lr_scheduler.CyclicLR(
+            optimizer,
+            base_lr=float(cfg.train_param.base_lr),
+            step_size_up=cfg.train_param.step_size_up,
+            max_lr=float(cfg.train_param.lr),
+            gamma=cfg.train_param.gamma,
+            mode='exp_range'
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'OneCycleLR':
+        scheduler = lr_scheduler.OneCycleLR(
+            optimizer,
+            max_lr=cfg.train_param.max_lr,
+            steps_per_epoch=cfg.train_param.steps_per_epoch,
+            epochs=cfg.train_param.epochs,
+            anneal_strategy='linear'
+        )
+    elif SCHEDULER_LIST[cfg.model_param.scheduler] == 'None':
+        return None
 
-
-def forward(
-    self,
-    input_ids=None,
-    attention_mask=None,
-    token_type_ids=None,
-    position_ids=None,
-    head_mask=None,
-    inputs_embeds=None,
-    labels=None,
-    output_attentions=None,
-    output_hidden_states=None,
-    return_dict=None,
-):
-    return_dict = return_dict if return_dict is not None else self.config.use_return_dict
-
-    # Funky alert
-    outputs = getattr(self, self.backbone_name)(
-        input_ids,
-        attention_mask=attention_mask,
-        token_type_ids=token_type_ids,
-        position_ids=position_ids,
-        # head_mask=head_mask, # these aren't necessary and some models error if you include
-        # inputs_embeds=inputs_embeds,  # these aren't necessary and some models error if you include
-        output_attentions=output_attentions,
-        output_hidden_states=output_hidden_states,
-        return_dict=return_dict,
-    )
-
-    sequence_output = outputs[0]
-
-    sequence_output = self.dropout(sequence_output)
-    logits = self.classifier(sequence_output)
-
-    loss = None
-    if labels is not None:
-        loss_fct = torch.nn.BCEWithLogitsLoss(reduction="none")
-        loss = loss_fct(logits.view(-1, 1), labels.view(-1, 1))
-
-        # this ignores the part of the sequence that got -100 as labels
-        loss = torch.masked_select(loss, labels.view(-1, 1) > -1).mean()
-
-    return TokenClassifierOutput(
-        loss=loss,
-        logits=logits,
-        hidden_states=outputs.hidden_states,
-        attentions=outputs.attentions,
-    )
-
-
-@classmethod
-def from_pretrained(cls, *args, **kwargs):
-    cls.PreTrainedModel = kwargs.pop("PreTrainedModel")
-    cls.ModelClass = kwargs.pop("ModelClass")
-    cls.backbone_name = kwargs["config"].model_type
-
-    # changes deberta-v2 --> deberta
-    if "deberta" in cls.backbone_name:
-        cls.backbone_name = "deberta"
-
-    return super(cls.PreTrainedModel, cls).from_pretrained(*args, **kwargs)
-
-
-def get_model(model_name_or_path, config):
-    model_type = type(config).__name__[:-len("config")]
-    name = f"{model_type}PreTrainedModel"
-    PreTrainedModel = getattr(__import__("transformers", fromlist=[name]),
-                              name)
-    name = f"{model_type}Model"
-    ModelClass = getattr(__import__("transformers", fromlist=[name]), name)
-
-    model = type('HybridModel', (PreTrainedModel,), {
-        '__init__': __init__,
-        "forward": forward,
-        "from_pretrained": from_pretrained
-    })
-
-    model._keys_to_ignore_on_load_unexpected = [r"pooler"]
-    model._keys_to_ignore_on_load_missing = [r"position_ids"]
-
-    return model.from_pretrained(
-        pretrained_model_name_or_path=model_name_or_path,
-        PreTrainedModel=PreTrainedModel,
-        ModelClass=ModelClass,
-        config=config
-    )
+    return scheduler
