+++ baseline2/medal_contender/preprocessing.py
@@ -2,7 +2,6 @@ import numpy as np
 import pandas as pd
 import ast
 
-
 # incorrect annotation
 def preprocessing_incorrect(train):
     train.loc[338, 'annotation'] = ast.literal_eval(
