+++ baseline3/medal_contender/preprocessing.py
@@ -2,6 +2,7 @@ import numpy as np
 import pandas as pd
 import ast
 
+
 # incorrect annotation
 def preprocessing_incorrect(train):
     train.loc[338, 'annotation'] = ast.literal_eval(
